move header and footer to libs
