+++ libs/layouts/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["../../.eslintrc.json"],
+  "ignorePatterns": ["!**/*"]
+}
