+++ libs/layouts/footer/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@ap/layouts/footer",
+  "sideEffects": false,
+  "type": "module"
+}
