+++ libs/layouts/footer/src/Newsletter/SubscriptionForm.tsx
