+++ libs/layouts/footer/src/Newsletter/index.ts
