+++ libs/layouts/footer/src/Newsletter/schema.ts
