+++ libs/layouts/footer/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "include": ["src/**/*.ts", "src/**/*.tsx"],
+  "extends": "../../../tsconfig.base.json"
+}
