+++ libs/layouts/header/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["plugin:@nrwl/nx/react", "../../../.eslintrc.json"],
+  "ignorePatterns": ["!**/*"]
+}
