+++ libs/layouts/header/project.json
@@ -1,0 +1,16 @@
+{
+  "name": "layouts-header",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/layouts/header/src",
+  "projectType": "library",
+  "tags": [],
+  "targets": {
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/layouts/header/**/*.{ts,tsx,js,jsx}"]
+      }
+    }
+  }
+}
