+++ libs/layouts/header/src/index.tsx
@@ -1,10 +1,10 @@
 import Logo, { DEFAULT_LOGO, LogoProps } from "@ap/components/logo";
 import WalletSuite from "@ap/components/wallet-suite";
 import { appRoutes } from "@ap/constants";
+import { Link } from "@ap/layouts";
 import { useEffect, useRef, useState } from "react";
 import { Location, matchRoutes, useLocation } from "react-router-dom";
-import { Link } from "../types";
-import Airdrop from "../Header/Airdrop";
+import Airdrop from "./Airdrop";
 import DesktopNav from "./DesktopNav";
 import { Opener as MobileNavOpener } from "./MobileNav";
 import ThemeToggle from "./ThemeToggle";
