+++ libs/layouts/project.json
@@ -1,0 +1,16 @@
+{
+  "name": "layouts",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/layouts/src",
+  "projectType": "library",
+  "targets": {
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/layouts/**/*.ts"]
+      }
+    }
+  },
+  "tags": []
+}
