+++ tsconfig.paths.json
@@ -38,6 +38,9 @@
 "@ap/hooks": ["libs/hooks/src/index.ts"],
 "@ap/hooks/use-cosmos-tx-sender": ["libs/hooks/use-cosmos-tx-sender/src/index.ts"],
 "@ap/hooks/use-sort": ["libs/hooks/use-sort/src/index.ts"],
+"@ap/layouts": ["libs/layouts/src/index.ts"],
+"@ap/layouts/footer": ["libs/layouts/footer/src/index.tsx"],
+"@ap/layouts/header": ["libs/layouts/header/src/index.tsx"],
 "@ap/schemas": ["libs/schemas/src/index.ts"],
 "@ap/services/apes": ["libs/services/apes/src/index.ts"],
 "@ap/services/aws": ["libs/services/aws/src/index.ts"],
